6 Size of

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("Size of a char: %lu byte(s)\n",(unsigned long)sizeof(a));
-printf("Size of an int: %lu byte(s)\n",(unsigned long)sizeof(b));
-printf("Size of a long int: %lu byte(s)\n",(unsigned long)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n",(unsigned long)sizeof(d));
-print("Size of a float: %lu byte(s)\n",(unsigned long)sizeof(f));
+printf("Size of a char: %zu byte(s)\n",sizeof(a));
+printf("Size of an int: %zu byte(s)\n",sizeof(b));
+printf("Size of a long int: %zu byte(s)\n",sizeof(c));
+printf("Size of a long long int: %zu byte(s)\n",sizeof(d));
+print("Size of a float: %zu byte(s)\n",sizeof(f));
 return (0);
 }
